pyproject.toml
[project]
name = "strands-async-tools"
version = "0.1.0"
description = "Async tool calling with callback-driven delivery for Strands Agents"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "strands-agents>=1.26.0",
    "textual>=7.5.0",
]
